python-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
groq==0.4.1
markdown==3.5
python-dotenv==1.0.0